Factor the text label rendering out of wset into a utility class.
This adds a simple node that can display text.
