+++ plugins/protocols/meson.build
@@ -3,7 +3,7 @@ protocol_plugins = [
 ]
 
 all_include_dirs = [wayfire_api_inc, wayfire_conf_inc, plugins_common_inc]
-all_deps = [wlroots, pixman, wfconfig, wf_protos, json]
+all_deps = [wlroots, pixman, wfconfig, wf_protos, json, cairo, pango]
 
 foreach plugin : protocol_plugins
   shared_module(plugin, plugin + '.cpp',
